refactor: extract PaperTradingPage chart + modal + formatters (F5)
Extract PortfolioGrowthChart (with its chart.js registration + Line import
and the portfolioTimeFrames data it owns) and TradeModal into
components/paper/, and the shared currency/number/percent formatters into
components/paper/format.js. PaperTradingPage keeps its portfolio state/
fetch logic, optimization data, and status helpers, dropping to 684 lines.
The chart.js register side-effect moves with the sole chart consumer.
Pure code movement, guarded by the existing test suite; lint + build clean.
